fix(atribuicao): conversões do Google sem arredondar (denominador da cobertura fiel)
Raiz do >100%: conversões do Google são fracionadas e o sync arredondava por dia
(0,4→0), deflacionando plataforma_compras. Agora:
- coluna atr_gastos.plataforma_compras: bigint -> numeric (ALTER, sem perda).
- ingest_gastos: cast ::numeric.
- sync google (.sh) + Edge Function: guardam conversões com 2 casas (não int).
- TabelaCampanhas: fmtC() exibe inteiro quando redondo, senão 1 casa.
Re-sync feito: 172 linhas voltaram a ter conversão fracionada real (antes 0/1).
Combinado com o teto de 100% (commit anterior), a cobertura fica correta.

+++ atribuicao/ingest_gastos.sql
@@ -32,7 +32,7 @@ BEGIN
       coalesce((r->>'gasto')::numeric, 0),
       coalesce((r->>'impressoes')::bigint, 0),
       coalesce((r->>'cliques')::bigint, 0),
-      coalesce((r->>'plataforma_compras')::bigint, 0),
+      coalesce((r->>'plataforma_compras')::numeric, 0),
       coalesce((r->>'plataforma_receita')::numeric, 0)
     )
     ON CONFLICT (dia, canal, coalesce(ad_id,''), coalesce(campaign_id,'')) DO UPDATE SET
